dump 浮点数提交。

+++ Assets/Script/Simulation/Components/MessageComponents.cs
@@ -5,6 +5,6 @@ using Unity.Mathematics;
 [Serializable]
 public struct MessageItem
 {
-    public float3 pos;
+    public int3 pos;
     public int id;
 }
